requirements.txt
Django>=2.2.0,<2.3.0
django-extensions>=2.2.8,<2.3.0
djangorestframework>=3.9.0,<3.10.0
django-rest-auth>=0.9.5,<0.10.0
django-allauth>=0.39.1,<0.40.0
psycopg2>=2.8.2,<2.9.0
flake8>=3.6.0,<3.7.0
Pillow>=5.3.0,<5.4.0
python-dotenv>=0.10.3,<0.11.0
pytest
pytest-django